feat: break N / continue N now match POSIX behavior
previously, providing a number to break/continue set the shell's status
now it breaks/continues out of N enclosing loops, as it does in bash

+++ src/builtin/hist.rs
@@ -256,12 +256,10 @@ impl HistQuery {
       entries.retain(|e| re.is_match(e.1.command()) != *not);
     }
 
-    if self.delete {
-      if !entries.is_empty() {
-        let ids: Vec<i64> = entries.iter().map(|e| e.0).collect();
-        hist.delete_ids(&ids)?;
-        hist.refresh_hist_entries();
-      }
+    if self.delete && !entries.is_empty() {
+      let ids: Vec<i64> = entries.iter().map(|e| e.0).collect();
+      hist.delete_ids(&ids)?;
+      hist.refresh_hist_entries();
     }
 
     // 'self.reverse' means 'print the entries in descending order'
